reengineered the training class to store NFE and average NFE
